new try

+++ hashmap.c
@@ -42,7 +42,7 @@ int is_equal(void* key1, void* key2){
 void insertMap(HashMap * map, char * key, void * value) {
   Pair* buckets=createPair(key,value)
 
-  int pos=hash(key,map->capacity);
+  long pos=hash(key,map->capacity);
   
   map->buckets[pos]=buckets;
   map->size++;
